Fix Net6 on linking when client nonce is shorter than 64 bytes.
Assuming that the client nonce message in the DH negotiation is exactly 64 bytes long is incorrect and causes the client to occasionally quit with "Network error 6, Disconnected From Server. protocol: Game Server" when linking. The message can be shorter because the client only sends the significant bits of the 512-bit number, omitting highest-order bytes when they are zero.

--HG--
branch : net6nonce