moss_uuidgen incorrectly includes the string terminator in the uuid.
The code sets the size of the variable it's creating (the uuid string)
to 37. This includes the nul string terminator. For unknown reasons
(newer PostgreSQL version?) this has started to cause UUID strings to
not match but look like they do (because the nul doesn't print), so ages
created by the server (such as during player creation) may not work.

--HG--
branch : de-bitrot-2018