Address the problem requiring standard_conforming_strings=off in PostgreSQL.
As of PostgreSQL 9.1 the default is conforming strings, while previous
versions defaulted to non-conforming. It appears that PostgreSQL's version
of binary string quoting took E'\x to mean "all the rest of this is hex",
so the string E'\x12345678' resulted in the DB containing four bytes
0x12 0x34 0x56 0x78. This unusual syntax is definitely not SQL standard --
with conforming strings, this is quite reasonably interpreted as one byte
in hex followed by a normal string, so 0x12 '3' '4' '5' '6' '7' '8'.

However, mind-bogglingly, the documentation does not appear to include a
function provided to quote binary data the conforming way, so this change
introduces one. It produces the same format as used by moss.sql (every byte
represented as octal). This quoting should be safe as it does not try to keep
"safe" characters represented as themselves. Now MOSS will work with either
setting of standard_conforming_strings because it no longer uses the weird
quoting.

--HG--
branch : de-bitrot-2018